requirements.txt
flask>=3.0
python-docx>=1.0
mammoth>=1.6
google-generativeai>=0.3
python-dotenv>=1.0
werkzeug>=3.0
gunicorn>=21.0
pytest>=7.0

